+++ studio/src/app/components/core/nav/app-navigation-actions/app-navigation-actions.scss
