+++ studio/src/app/components/core/nav/app-navigation/app-navigation.scss
